src/requirements.txt
gunicorn
python-levenshtein
flask
redis